requirements.txt
python-pdfbox
beautifulsoup4
fake-useragent
selenium
pypdf
huggingface_hub
acryl-datahub[datahub-rest]
datahub
torch
transformers
python-dotenv
langchain
tiktoken
webdriver-manager
mlflow==2.4.1
greenplum-python
torchvision
accelerate

matplotlib
pmdarima
prophet
psycopg2-binary
sqlalchemy
python-dotenv
pandas
wheel
scikit-learn
nbconvert
seaborn
traitlets
scipy
plotly
watchdog
mpld3
streamlit==1.18.0
Pillow
streamlit-autorefresh
jupyterflow
lz4
pyyaml
prometheus-client
protobuf # protobuf==3.20.0 # protobuf==3.19.5
fastapi
uvicorn[standard]
gunicorn
boto3
tensorflow
rsocket
python-snappy
filelock==3.9.0
pyarrow<7.0.0
hickle
python-multipart
evidently
pysqlite3-binary
kfp==1.8.19
python-multipart
pyservicebinding
download_and_extract
altair==4.0
.